add level mode splendor, splendor_v2, sushigo, machikoro, sheriff, century
